Update Crud Nested Model

+++ nest/Controller.php
@@ -2,25 +2,30 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\BASE_Upper_singular;
 use App\Models\NEST_Upper_singular;
+use App\Models\BASE_Upper_singular;
 use App\Http\Requests\BASE_Upper_singularRequest;
 use App\Http\Resources\BASE_Upper_singularResource;
 
 class BASE_Upper_singularController extends Controller
 {
     /**
-     * Create the controller instance.
+     * Create a new BASE_lower_singular for the NEST_lower_singular.
      *
-     * @return void
+     * @param App\Models\NEST_Upper_singular                $NEST_lower_singular   The NEST_lower_singular
+     * @param App\Http\Requests\BASE_Upper_singularRequest $validated The validated request
+     *
+     * @return \Illuminate\Http\Response
      */
-    public function __construct()
+    public function store(NEST_Upper_singular $NEST_lower_singular, BASE_Upper_singularRequest $validated)
     {
-        $this->authorizeResource(BASE_Upper_singular::class, 'BASE_lower_singular');
+        $this->authorize('create', [new BASE_Upper_singular, $NEST_lower_singular]);
+
+        return new BASE_Upper_singularResource($NEST_lower_singular->BASE_lower_plural()->create($validated->toArray()));
     }
 
     /**
-     * Display a listing of the resource.
+     * Return all the BASE_lower_plural for the NEST_lower_singular.
      *
      * @param NEST_Upper_singular $NEST_lower_singular The NEST_lower_singular
      *
@@ -28,72 +33,64 @@ class BASE_Upper_singularController extends Controller
      */
     public function index(NEST_Upper_singular $NEST_lower_singular)
     {
-        if (true) { // TODO CRUD ADD AUTH CHECK HERE
-            return BASE_Upper_singularResource::collection($NEST_lower_singular->BASE_lower_plural);
-        }
+        $this->authorize('viewAny', [new BASE_Upper_singular, $NEST_lower_singular]);
 
-        return abort(403, "Unauthorized: You can not create a new BASE_lower_singular for this NEST_lower_singular");
+        return BASE_Upper_singularResource::collection($NEST_lower_singular->BASE_lower_plural);
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Return the specified BASE_lower_singular for the NEST_lower_singular.
      *
-     * @param NEST_Upper_singular    $NEST_lower_singular The NEST_lower_singular
-     * @param BASE_Upper_singularRequest $request The request
+     * @param String $NEST_lower_singularId  The NEST_lower_singular id
+     * @param String $BASE_lower_singularId The NEST_lower_singular id
      *
      * @return \Illuminate\Http\Response
      */
-    public function store(NEST_Upper_singular $NEST_lower_singular, BASE_Upper_singularRequest $request)
+    public function show(String $NEST_lower_singularId, String $BASE_lower_singularId)
     {
-        if (true) { // TODO CRUD ADD AUTH CHECK HERE
-            return new BASE_Upper_singularResource($NEST_lower_singular->BASE_lower_plural()->create($request->validated()));
-        }
+        $this->authorize('view', [BASE_Upper_singular::find($BASE_lower_singularId), NEST_Upper_singular::find($NEST_lower_singularId)]);
 
-        return abort(403, "Unauthorized: You can not create a new BASE_lower_singular for this NEST_lower_singular");
+        return new BASE_Upper_singularResource(BASE_Upper_singular::find($BASE_lower_singularId));
     }
 
     /**
-     * Display the specified resource.
+     * Update the specified BASE_lower_singular for the NEST_lower_singular.
      *
-     * @param \App\Models\BASE_lower_singular $BASE_lower_singular The BASE_lower_singular
+     * @param int                               $NEST_lower_singularId  The requested NEST_lower_singular id
+     * @param int                               $BASE_lower_singularId The requested BASE_lower_singular id
+     * @param App\Http\Requests\BASE_Upper_singularRequest $validated  The validated request
      *
      * @return \Illuminate\Http\Response
      */
-    public function show(BASE_Upper_singular $BASE_lower_singular)
+    public function update($NEST_lower_singularId, $BASE_lower_singularId, BASE_Upper_singularRequest $validated)
     {
-        return new BASE_Upper_singularResource($BASE_lower_singular);
+        $this->authorize('update', [BASE_Upper_singular::find($BASE_lower_singularId), NEST_Upper_singular::find($NEST_lower_singularId)]);
+
+        BASE_Upper_singular::find($BASE_lower_singularId)->update($validated->toArray());
+
+        return new BASE_Upper_singularResource(BASE_Upper_singular::find($BASE_lower_singularId));
     }
 
     /**
-     * Update the specified resource in storage.
+     * Delete the specified BASE_lower_singular for the NEST_lower_singular.
      *
-     * @param BASE_Upper_singularRequest $request The request
-     * @param BASE_Upper_singular        $BASE_lower_singular     The BASE_lower_singular
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function update(BASE_Upper_singularRequest $request, BASE_Upper_singular $BASE_lower_singular)
-    {
-        $BASE_lower_singular->update($request->validated());
-
-        return new BASE_Upper_singularResource($BASE_lower_singular);
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param \App\Models\BASE_Upper_singular $BASE_lower_singular The BASE_lower_singular
+     * @param int $NEST_lower_singularId  The requested NEST_lower_singular id
+     * @param int $BASE_lower_singularId The requested BASE_lower_singular id
      *
      * @return \Illuminate\Http\Response
      */
-    public function destroy(BASE_Upper_singular $BASE_lower_singular)
+    public function destroy($NEST_lower_singularId, $BASE_lower_singularId)
     {
+        $BASE_lower_singular = BASE_Upper_singular::find($BASE_lower_singularId);
+
+        $this->authorize('delete', [$BASE_lower_singular, NEST_Upper_singular::find($NEST_lower_singularId)]);
+
         $BASE_lower_singular->delete();
 
         return response()->json(
             [
                 'message' => 'BASE_Upper_singular deleted.',
-                'NEST_lower_singular' => new BASE_Upper_singularResource($BASE_lower_singular)
+                'BASE_lower_singular' => new BASE_Upper_singularResource($BASE_lower_singular)
             ]
         );
     }
